Reorganize QueryNodeEditor component
