+++ frontend/lib/styles/components/queryNodeEditor.sass
@@ -77,6 +77,9 @@
     font-size: $font-md
     font-weight: 700
     color: $col-black
+    > p
+      margin: 0
+      line-height: inherit
 
   .input--value-changed
     .Select
